% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunTools.R
\name{calcuPairSpecSim}
\alias{calcuPairSpecSim}
\title{calcuPairSpecSim}
\usage{
calcuPairSpecSim(
  annot_a = "C00204",
  annot_a_peak = c("M177T553"),
  annot_b = "C04349",
  annot_b_peak = c("M235T596", "M157T548"),
  ms2_data = raw_msms
)
}
\description{
calcuPairSpecSim
}
\author{
Zhiwei Zhou
}
