% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{checkFiles4Vis}
\alias{checkFiles4Vis}
\title{checkFiles4Vis}
\usage{
checkFiles4Vis(dir_path = ".")
}
\arguments{
\item{dir_path}{'.'}
}
\description{
Check required files for visualization
}
\author{
Zhiwei Zhou
}
