% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{copyFiles4Vis}
\alias{copyFiles4Vis}
\title{copyFiles4Vis}
\usage{
copyFiles4Vis(dir_path = ".")
}
\arguments{
\item{dir_path}{the path of KGMN result folder}
}
\description{
copy intermidate data files for visualization
}
\author{
Zhiwei Zhou
}
