% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{extractNetwork2FromNode}
\alias{extractNetwork2FromNode}
\title{extractNetwork2FromNode}
\usage{
extractNetwork2FromNode(
  id_kegg,
  peak,
  annotation_table,
  network_emrn,
  ms2,
  reaction_step = 1,
  ms2_link = c("smilarity", "hybrid")
)
}
\description{
extractNetwork2FromNode
}
\author{
Zhiwei Zhou
}
