% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{getCompoundInfo}
\alias{getCompoundInfo}
\title{getCompoundInfo}
\usage{
getCompoundInfo(query_id)
}
\arguments{
\item{query_id}{The compound id in annotation_table one. Single compound query is supported}
}
\description{
query compound information in KGMN
}
\examples{
# compounds in KMRN
getCompoundInfo('C01802')
getCompoundInfo('KeggExd000923')

# compounds in library
getCompoundInfo('L0449')
}
\author{
Zhiwei Zhou
}
