% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{mergeSubnetwork}
\alias{mergeSubnetwork}
\title{mergeSubnetwork}
\usage{
mergeSubnetwork(
  from_peak,
  end_peak,
  dir_path = ".",
  folder_output = NULL,
  show_plot = TRUE
)
}
\arguments{
\item{dir_path}{path of working directory. Default: '.'}

\item{folder_output}{path of subnetwork export. It will generate a subfolder in 03_subnetwork folder. If no folder defined, the subnetworks will exported into a folder with name of centric_met. Default: NULL}

\item{show_plot}{whether show merged network. Default: TRUE}
}
\description{
merge subnetwork2 and subnetwork3 from KGMN. Note: this function only effective after performing subnetwork2 and subnetwork3 extraction
}
\examples{
mergeSubnetwork(from_peak = 'M262T526', end_peak = 'M182T541', dir_path = 'D:/project/00_zhulab/01_metdna2/00_data/20220602_visualization_kgmn/06_visualization/')
}
\author{
Zhiwei Zhou
}
