% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{reconstructNetwork2}
\alias{reconstructNetwork2}
\title{reconstructNetwork2}
\usage{
reconstructNetwork2(
  annotation_table,
  dir_path = ".",
  is_unknown_annotation = TRUE,
  whether_show_single_node = FALSE,
  max_reaction_step = 3,
  ms2_link = c("hybrid", "smilarity")
)
}
\arguments{
\item{whether_show_single_node}{Default: FALSE}

\item{max_reaction_step}{Default: 3}

\item{ms2_link}{'hybrid' or 'smilarity'; 'hybrid' means linking nodes through MS/MS similarity (>= 0.5) or matched fragments (>=4); 'smilarity' means linking nodes through MS/MS similarity (>= 0.5) only; Default: hybrid}
}
\description{
reconstructNetwork2
}
\author{
Zhiwei Zhou
}
