% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{reconstructNetwork3}
\alias{reconstructNetwork3}
\title{reconstructNetwork3}
\usage{
reconstructNetwork3(dir_path = ".")
}
\arguments{
\item{dir_path}{path of working directory}
}
\value{
a list of network3, includes 2 objects (node_table & edge_table)
}
\description{
retrieve network3 for visualization
}
\examples{
setwd('D:/project/00_zhulab/01_metdna2/00_data/20220602_visualization_kgmn/06_visualization')
reconstructNetwork3()
}
\author{
Zhiwei Zhou
}
