% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{retrieveSubNetwork1}
\alias{retrieveSubNetwork1}
\title{retrieveSubNetwork1}
\usage{
retrieveSubNetwork1(
  centric_met,
  step = 1,
  is_unknown_annotation = TRUE,
  dir_path = ".",
  folder_output = NULL,
  show_plot = TRUE,
  mode = c("emrn", "46std")
)
}
\arguments{
\item{centric_met}{charecter vector. If the length of centric_met equals 1, then retrieve subnetwork with defined step. If the length of centric_met >= 2, then only retrieve subnetwork with nodes from input centric_met.}

\item{step}{neigbor step. Default: 1}

\item{is_unknown_annotation}{whether used unknown annotation. Default: TRUE}

\item{dir_path}{path of working directory. Default: '.'}

\item{folder_output}{path of subnetwork export. It will generate a subfolder in 03_subnetwork folder. If no folder defined, the subnetworks will exported into a folder with name of centric_met. Default: NULL}

\item{show_plot}{whether show subnetwork plot in R. Default: TRUE}

\item{mode}{'emrn' and '46std'. Default: "emrn". The "46std" is only used in in-vitro metabolism experiment.}
}
\description{
retrieve subnetwork1 from constructed metabolic reaction network
}
\examples{
# single node input
retrieveSubNetwork1()
}
