% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{retrieveSubNetwork2}
\alias{retrieveSubNetwork2}
\title{retrieveSubNetwork2}
\usage{
retrieveSubNetwork2(
  from_peak,
  end_peak = NULL,
  step = 1,
  dir_path = ".",
  folder_output = NULL,
  show_plot = TRUE
)
}
\arguments{
\item{end_peak}{Character. Default: NULL}

\item{step}{Numeric. Default: step = 1}

\item{dir_path}{path of working directory. Default: '.'}

\item{folder_output}{path of subnetwork export. It will generate a subfolder in 03_subnetwork folder. If no folder defined, the subnetworks will exported into a folder with name of centric_met. Default: NULL}
}
\description{
retrieveSubNetwork2
}
\examples{
# subnetwork from centric peak
  retrieveSubNetwork2(from_peak = 'M196T420', step = 1, dir_path = 'D:/project/00_zhulab/01_metdna2/00_data/20220602_visualization_kgmn/06_visualization')

# subnetwork with defined from_peak and end_peak
   retrieveSubNetwork2(from_peak = 'M262T526', end_peak = 'M182T541', dir_path = 'D:/project/00_zhulab/01_metdna2/00_data/20220602_visualization_kgmn/06_visualization')

}
\author{
Zhiwei Zhou
}
