% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{retrieveSubNetwork3}
\alias{retrieveSubNetwork3}
\title{retrieveSubNetwork3}
\usage{
retrieveSubNetwork3(
  base_peaks,
  base_adducts,
  dir_path = ".",
  folder_output = NULL,
  show_plot = TRUE
)
}
\arguments{
\item{base_peaks}{a vector of base_peaks}

\item{base_adducts}{a vector of base_adducts. Note: it should keep same length with base_peaks}

\item{dir_path}{path of working directory.}

\item{folder_output}{path of subnetwork export. It will generate a subfolder in 03_subnetwork folder. If no folder defined, the subnetworks will exported into a folder with name of centric_met. Default: NULL}

\item{show_plot}{Whether show subnetwork plot. Default: TRUE}
}
\description{
retrieveSubNetwork3
}
\examples{
retrieveSubNetwork3(base_peaks = c('M182T541', 'M262T526'),
base_adducts = c('[M+H]+', '[M+H]+'))
}
\author{
Zhiwei Zhou
}
