% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunVisualization.R
\name{reconstructNetwork1}
\alias{reconstructNetwork1}
\title{reconstructNetwork1}
\usage{
reconstructNetwork1(dir_path = ".", is_unknown_annotation = TRUE)
}
\arguments{
\item{dir_path}{path of working directory. Default: '.'}

\item{is_unknown_annotation}{whether used unknown annotation. Default: TRUE}
}
\description{
reconstructNetwork1
}
\author{
Zhiwei Zhou
}
